Renamed & updated JDBC example
